Create front individual analysis route
