+++ orga/app/styles/app.scss
@@ -45,7 +45,9 @@
 @import "pages/authenticated/campaigns/details/parameters";
 @import "pages/authenticated/campaigns/details/collective-results/success-indicator";
 @import "pages/authenticated/campaigns/details/participants";
-@import "pages/authenticated/campaigns/details/participants/participant-results";
+@import "pages/authenticated/campaigns/details/participants/participant";
+@import "pages/authenticated/campaigns/details/participants/participant/header";
+@import "pages/authenticated/campaigns/details/participants/participant/results";
 @import "pages/authenticated/team";
 @import "pages/authenticated/team/list";
 @import "pages/authenticated/team/new";
