--- orga/app/styles/pages/authenticated/campaigns/details/participants/participant-results.scss
@@ -1,136 +1,0 @@
-.participant-results {
-  width: 100%;
-
-  &__header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin: 16px 0;
-  }
-
-  &__information {
-    margin-bottom: 10px;
-
-  }
-
-  &__result {
-    padding: 10px;
-  }
-
-  &__details {
-    margin-bottom: 30px;
-  }
-
-  &-row {
-    margin-bottom: 1em;
-  }
-}
-
-.participant-results-header {
-  &__title {
-    display: flex;
-    align-items: center;
-  }
-}
-
-.participant-results-row {
-  display: flex;
-  justify-content: space-between;
-}
-
-.participant-results-content {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  margin: 0 2%;
-
-  &--large {
-    padding: 30px 0;
-  }
-
-  &--wide {
-    width: 100%;
-  }
-
-  &__left-wrapper {
-    border-right: 1px solid $grey-20;
-    width: 60%;
-    margin: 0 0 0 3%;
-    flex-shrink: 1;
-  }
-
-  &__right-wrapper {
-    display: flex;
-    align-items: flex-start;
-    justify-content: space-around;
-    flex-shrink: 0;
-    flex-grow: 1;
-    min-width: 370px;
-    max-width: 600px;
-
-    .participant-results-content--large {
-      flex-shrink: 0;
-    }
-  }
-
-  &__summary {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    width: 50%;
-    padding: 0 30px;
-  }
-
-  &__progress-bar {
-    position: relative;
-  }
-
-  &__score {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 240px;
-  }
-
-  &__circle-chart {
-    margin-right: 40px;
-  }
-
-  &__circle-chart-value {
-    color: $grey-100;
-    font-family: $roboto;
-    font-size: 1.9375rem;
-  }
-
-  &__label-text {
-    margin-bottom: 10px;
-  }
-
-  &__check-shared {
-    margin-right: 3px;
-
-    &--green {
-      color: $success;
-    }
-  }
-}
-
-.participant-results-report {
-  display: flex;
-  align-items: center;
-  width: 80px;
-  height: 80px;
-  text-align: center;
-
-  &__percent {
-    position: absolute;
-    text-align: center;
-  }
-
-  &__percent-text {
-    color: $grey-60;
-    font-family: $roboto;
-    font-size: 2rem;
-    padding: 10px;
-  }
-}
