+++ orga/app/styles/pages/authenticated/campaigns/details/participants/participant/results.scss
@@ -1,0 +1,64 @@
+.participant-results {
+  width: 100%;
+
+  &__result {
+    padding: 10px;
+  }
+
+  &__details {
+    margin-bottom: 30px;
+  }
+}
+
+.participant-row {
+  &__result {
+    margin-bottom: 1em;
+  }
+}
+
+.participant-results-content {
+  &__summary {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    width: 50%;
+    padding: 0 30px;
+  }
+
+  &__score {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    width: 240px;
+  }
+
+  &__circle-chart {
+    margin-right: 40px;
+  }
+
+  &__circle-chart-value {
+    color: $grey-100;
+    font-family: $roboto;
+    font-size: 1.9375rem;
+  }
+}
+
+.participant-results-report {
+  display: flex;
+  align-items: center;
+  width: 80px;
+  height: 80px;
+  text-align: center;
+
+  &__percent {
+    position: absolute;
+    text-align: center;
+  }
+
+  &__percent-text {
+    color: $grey-60;
+    font-family: $roboto;
+    font-size: 2rem;
+    padding: 10px;
+  }
+}
