+++ orga/app/templates/authenticated/campaigns/details/participants/participant.hbs
@@ -1,0 +1,6 @@
+<div class="participant">
+  <Routes::Authenticated::Campaigns::Details::Participants::ParticipantHeader
+          @campaignParticipation={{@model.campaignParticipation}}
+          @campaign={{@model.campaign}} />
+  {{outlet}}
+</div>
