+++ orga/app/templates/authenticated/campaigns/details/participants/participant/analysis.hbs
@@ -1,0 +1,2 @@
+<Routes::Authenticated::Campaigns::Details::AnalysisTab
+  @campaignTubeRecommendations={{@model.campaignParticipation.campaignAnalysis.campaignTubeRecommendations}}/>
