+++ orga/app/templates/authenticated/campaigns/details/participants/participant/results.hbs
@@ -1,0 +1,1 @@
+<Routes::Authenticated::Campaigns::Details::Participants::ResultsItem @campaignParticipation={{@model.campaignParticipation}} @class="participant-results" />
