--- orga/app/templates/authenticated/campaigns/details/participants/results.hbs
@@ -1,1 +1,0 @@
-<Routes::Authenticated::Campaigns::Details::Participants::ResultsItem @campaignParticipation={{@model.campaignParticipation}} @campaign={{@model.campaign}} @class="participant-results" />
