+++ orga/app/templates/components/routes/authenticated/campaigns/details/participants/results-item.hbs
@@ -1,72 +1,14 @@
-<div class="participant-results__header">
-  <div class="participant-results-header__title">
-    <LinkTo @route="authenticated.campaigns.details.participants" @model={{@campaign.id}} class="icon-button campaign-details-content__return-button">
-      <FaIcon @icon='arrow-left'></FaIcon>
-    </LinkTo>
-    <div
-      class="page__title page-title">{{@campaignParticipation.user.firstName}} {{@campaignParticipation.user.lastName}}</div>
-  </div>
-</div>
-
-
-<div class="panel panel--light-shadow participant-results__information">
-  <div class="participant-results-row">
-    {{#if @campaignParticipation.participantExternalId}}
-      <div
-        class="participant-results-content participant-results-content--large participant-results-content__left-wrapper">
-        <div class="label-text participant-results-content__label-text">{{@campaign.idPixLabel}}</div>
-        <div class="content-text">{{@campaignParticipation.participantExternalId}}</div>
-      </div>
-    {{/if}}
-    <div class="participant-results-content__right-wrapper">
-      <div class="participant-results-content participant-results-content--large">
-        <div class="label-text participant-results-content__label-text">Avancement</div>
-        <div aria-label="Avancement" class="content-text">
-          {{percentage @campaignParticipation.campaignParticipationResult.progress}}%
-          <div class="participant-results-content__progress-bar progress-bar">
-            {{#if @campaignParticipation.campaignParticipationResult.isCompleted}}
-              <div class="participant-results-content__progress-bar progress-bar progress-bar--completed"></div>
-            {{else}}
-              <div class="progress-bar progress-bar--completion" style={{@progressBarStyle}}>
-              </div>
-            {{/if}}
-          </div>
-        </div>
-      </div>
-      <div class="participant-results-content participant-results-content--large">
-        <div class="label-text participant-results-content__label-text">Commencé le</div>
-        <div class="content-text">
-          {{moment-format @campaignParticipation.createdAt 'll'}}
-        </div>
-      </div>
-      <div class="participant-results-content participant-results-content--large">
-        <div class="label-text participant-results-content__label-text">Partagé le</div>
-        <div class="content-text">
-          {{#if @campaignParticipation.isShared}}
-            <FaIcon @icon='check-circle'
-                      @class="participant-results-content__check-shared participant-results-content__check-shared--green"></FaIcon>
-            {{moment-format @campaignParticipation.sharedAt 'll'}}
-          {{else}}
-            Non disponible
-          {{/if}}
-
-        </div>
-      </div>
-    </div>
-
-  </div>
-</div>
-<div class="panel participant-results-row">
+<div class="panel participant-row participant-row__result">
   <div class="participant-results-content__summary">
-    <div class="participant-results-content participant-results-content--large participant-results-content--wide">
-      <div class="label-text participant-results-content__label-text">Acquis validés</div>
+    <div class="participant-content participant-content--large participant-content--wide">
+      <div class="label-text participant-content__label-text">Acquis validés</div>
       <div class="content-text content-text--big">
         {{if @campaignParticipation.isShared
              @campaignParticipation.campaignParticipationResult.validatedSkillsCount '-' }}
       </div>
     </div>
-    <div class="participant-results-content participant-results-content--large participant-results-content--wide">
-      <div class="label-text participant-results-content__label-text">Acquis évalués</div>
+    <div class="participant-content participant-content--large participant-content--wide">
+      <div class="label-text participant-content__label-text">Acquis évalués</div>
       <div class="content-text content-text--big">
         {{if @campaignParticipation.isShared
              @campaignParticipation.campaignParticipationResult.totalSkillsCount '-' }}
@@ -74,12 +16,12 @@
     </div>
   </div>
 
-  <div class="participant-results-content">
-    <div class="participant-results-content__score">
+  <div class="participant-content participant-results-content">
+    <div class=" participant-results-content__score">
       <div class="content-text content-text--big content-text--bold">Résultat</div>
-      <div class="participant-results-content__circle-chart">
+      <div class=" participant-results-content__circle-chart">
         <CircleChart @value={{@campaignParticipation.campaignParticipationResult.masteryPercentage}} @isDisabled={{not @campaignParticipation.isShared}}>
-          <span class="participant-results-content__circle-chart-value">
+          <span class=" participant-results-content__circle-chart-value">
             {{@campaignParticipation.campaignParticipationResult.masteryPercentage}}%
           </span>
         </CircleChart>
